+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.2.3] - 2026-09-24
+## [0.2.3] - 2026-09-23
 
 ### Added
 - **DNS zone creation and deletion** on the DNS page